Fix fullscreen crash.
If fullscreen is set programmatically, PPAPI fails -- but LOVE asserts
that it was able to set the graphics mode properly.

This change makes it always succeed, even if setting fullscreen fails.

Also added a NULL pointer check to FixEvent -- it's possible to receive
events before DidChangeView is called (but that is a separate issue).

Also leak a reference to the Window -- we don't ever want it destroyed;
this caused a crash on exit.