Remove calls to glGetError in VBO.
glGetError is very slow on NaCl. Some games create many VBOs on the
fly, and each one would call glGetError at least twice.

We now just assume the call succeeds; any errors will be displayed in
the Chrome JavaScript console, and it is unlikely that any games depend
on the behavior of having a VBO creation fail.

This fixes issue #31.